% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{add_scans}
\alias{add_scans}
\title{TO WRITE}
\usage{
add_scans(scan.list, new.scans)
}
\arguments{
\item{scan.list}{TO WRITE}

\item{new.scans}{}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
