% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{add_scans.scanList}
\alias{add_scans.scanList}
\title{TO WRITE}
\usage{
\method{add_scans}{scanList}(scan.list, new.scans)
}
\arguments{
\item{scan.list}{TO WRITE}

\item{new.scans}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
