% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_mode_tools.R
\name{apply_mode}
\alias{apply_mode}
\title{Make Adjacency fit the selected mode
Internal use.}
\usage{
apply_mode(
  raw.scanList,
  mode = c("directed", "undirected", "max", "min", "upper", "lower", "plus", "vector")
)
}
\arguments{
\item{raw.scanList}{a list of (directed by default) binary adjacency matrix, pre-theoretical, to
which the chosen \code{mode} is to be applied}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. Default
here is directed. Possible values are: directed, undirected, upper, lower, max, min, plus.
Added vector too. See details
\code{\link[igraph:graph_from_adjacency_matrix]{graph_from_adjacency_matrix()}}.}
}
\value{
an list of adjacency matrix fitting the chosen \code{mode} is to be applied
}
\description{
Make Adjacency fit the selected mode
Internal use.
}
\keyword{internal}
