% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_tools.R
\name{attrs<-}
\alias{attrs<-}
\title{TO WRITE}
\usage{
attrs(x, which) <- value
}
\arguments{
\item{x}{TO WRITE}

\item{which}{TO WRITE}

\item{value}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
