% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_tools.R
\name{attrs}
\alias{attrs}
\title{TO WRITE}
\usage{
attrs(scanList, a = NULL)
}
\arguments{
\item{scanList}{TO WRITE}

\item{a}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
