% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.tools.R
\name{binary_adjacency_mode}
\alias{binary_adjacency_mode}
\title{Make Binary Adjacency fit the selected mode
From a directed binary adjacency matrix, make it fit the selected mode.}
\usage{
binary_adjacency_mode(
  Adj,
  mode = c("directed", "undirected", "max", "min", "upper", "lower", "plus")
)
}
\arguments{
\item{Adj}{a binary adjacency matrix}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.}
}
\value{
a binary adjacency matrix fitting the selected mode
}
\description{
Make Binary Adjacency fit the selected mode
From a directed binary adjacency matrix, make it fit the selected mode.
}
\examples{
set.seed(42)

n<- 5;nodes<- letters[1:n];
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[non.diagonal(Adj)]<- sample(0:30,n*(n-1),replace = TRUE)
Adj

Adj<- do.scan(Adj,42)
binary_adjacency_mode(Adj$theoretical,"plus")
}
