% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Boot_get.list}
\alias{Boot_get.list}
\title{Retrieve data from specific method from Boot_scans() output
Subset rich Bootstrap output choosing what's needed}
\usage{
Boot_get.list(
  Bootstrap,
  what = c("theoretical", "group", "focal"),
  get.format = c("list", "adjacency", "observed_edges", "all")
)
}
\arguments{
\item{Bootstrap}{Bootstrap output object}

\item{what}{character scalar, data type requested ("theoretical","group" or "focal")}

\item{get.format}{character scalar, output format type requested ("list","adjacency", "observed_edges" or "all").}
}
\value{
list which structure depends on chosen data type and Bootstrap attributes
}
\description{
Retrieve data from specific method from Boot_scans() output
Subset rich Bootstrap output choosing what's needed
}
\examples{

set.seed(42)

n<- 5;nodes<- letters[1:n];
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[non.diagonal(Adj)]<- sample(0:42,n*(n-1),replace = TRUE)
Adj

focal.list<- sample(nodes,42,replace = TRUE)
Bootstrap<- Boot_scans(Adj,3,total_scan = 42,focal.list = focal.list,
                       scaled = FALSE,obs.prob=0.7,
                       method = "group",mode = "directed",output = "list")
Boot_get.list(Bootstrap,"theoretical")
Boot_get.list(Bootstrap,"group")
Boot_get.list(Bootstrap,"group","adj")
Boot_get.list(Bootstrap,"group","obs")

Bootstrap<- Boot_scans(Adj,3,total_scan = 42,focal.list = focal.list,
                       scaled = FALSE,obs.prob=0.7,
                       method = "group",mode = "directed",output = "adjacency")
Boot_get.list(Bootstrap,"theoretical","adj")
# Boot_get.list(Bootstrap,"group","list")
Boot_get.list(Bootstrap,"group","adj")
Boot_get.list(Bootstrap,"group","obs")

Bootstrap<- Boot_scans(Adj,3,total_scan = 42,focal.list = focal.list,
                       scaled = TRUE,obs.prob=0.7,keep=TRUE,
                       method = "both",mode = "directed",output = "all")
Boot_get.list(Bootstrap,"focal","all")
Boot_get.list(Bootstrap,"group","adjacency")
Boot_get.list(Bootstrap,"focal","obs")
Boot_get.list(Bootstrap,"group","list")
}
