% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Bootstrap.list_post.scale}
\alias{Bootstrap.list_post.scale}
\title{Scale unscaled bootstrap list}
\usage{
Bootstrap.list_post.scale(Bootstrap)
}
\arguments{
\item{Bootstrap}{Bootstrap output object (/!\ WITH output = "adjacency", other output to be implemented if relevant)}
}
\value{
a Bootstrap output object with scaled adjacencies (comparable )
}
\description{
Scale unscaled bootstrap list
}
\examples{
# Internal use.
}
