% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Bootstrap_add.attributes}
\alias{Bootstrap_add.attributes}
\title{Keep bootstrap parameters (and more) as attributes
Internal use in Boot_scans(). Add "method", "keep", "mode", "output" attributes to be more easily retrieved by the get function}
\usage{
Bootstrap_add.attributes(Bootstrap, ...)
}
\arguments{
\item{Bootstrap}{Boot_scans() intermediate output}

\item{...}{Any named attribute to be included to the Bootstrap object. Mostly determined by the code of Boot_scan(). Will produce error later if no name is attributed to any}
}
\value{
Bootstrap object, with stored attribute for later retrieval through Bootstrap_get.attributes(). A special attribute, attr.list stores those who have been properly inputted, to account for those with special structure.
}
\description{
Keep bootstrap parameters (and more) as attributes
Internal use in Boot_scans(). Add "method", "keep", "mode", "output" attributes to be more easily retrieved by the get function
}
\examples{
#Internal
}
