% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Bootstrap_get.attr}
\alias{Bootstrap_get.attr}
\title{Retrieve bootstrap object's attributes and reassign them in caller frame}
\usage{
Bootstrap_get.attr(Bootstrap, a)
}
\arguments{
\item{Bootstrap}{Boot_scans() output}

\item{a}{character, parameter stored as attribute to retrieve from `Bootstrap`.}
}
\value{
the value of the parameter `a`.
}
\description{
Retrieve bootstrap object's attributes and reassign them in caller frame
}
\examples{
#Internal
}
