% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{cbind_lapply}
\alias{cbind_lapply}
\title{Column bind list of data frames
wrapper to one-function do.call cbind over a lapply list}
\usage{
cbind_lapply(X, FUN)
}
\arguments{
\item{X}{a list. See details \link[base]{lapply}.}

\item{FUN}{a function to subset data frames (or data tables). See details \link[base]{lapply}.}
}
\value{
a column bound data frame
}
\description{
Column bind list of data frames
wrapper to one-function do.call cbind over a lapply list
}
\examples{
set.seed(42)

X<- lapply(1:3,function(i) list(int = 42,df = data.frame(x = runif(10,0,1),y = runif(10,0,1))))
cbind_lapply(X,function(x) x$df)
}
