% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_sampling.R
\name{check_sampling_parameters}
\alias{check_sampling_parameters}
\title{TO WRITE}
\usage{
check_sampling_parameters(
  method,
  sampling = c("constant", "matrix", "even", "random", "function")
)
}
\arguments{
\item{method}{TO WRITE}

\item{sampling}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\keyword{internal}
