% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.scan.R
\name{comp_with_transposed}
\alias{comp_with_transposed}
\title{Compare elements of a matrix with its transposed
wrapper using `null_na`}
\usage{
comp_with_transposed(X, comp.fun = `>=`)
}
\arguments{
\item{X}{a numeric matrix}

\item{comp.fun}{a function to compare X with t(X), in this order. Default is superior or equal}
}
\value{
a logical matrix
}
\description{
Compare elements of a matrix with its transposed
wrapper using `null_na`
}
\examples{
# Internal use
}
