% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{compute.strength}
\alias{compute.strength}
\title{Compute node strength from graph or adjacency matrix}
\usage{
compute.strength(graph, mode = NULL)
}
\arguments{
\item{graph}{an igraph object (or an adjacency matrix)}

\item{mode}{optional, only if `graph` is an adjacency matrix. Otherwise character scalar,
specifies how igraph should interpret the supplied matrix. Default here is directed. Possible
values are: directed, undirected, upper, lower, max, min, plus. Added vector too. See details
\link[igraph]{graph_from_adjacency_matrix}.}
}
\value{
a vector of strength values for each node
}
\description{
Compute node strength from graph or adjacency matrix
}
\examples{
set.seed(42)

n<- 5;nodes<- letters[1:n];total_scan<- 42;
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[non.diagonal(Adj)]<- sample(0:total_scan,n*(n-1),replace = TRUE)
Adj

compute.strength(Adj,"directed")
}
