% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANTs_tools.R
\name{convert_scan_to_df}
\alias{convert_scan_to_df}
\title{Convert a scan object into a data frame
Helper function to ensure compatibility with the ANTs package}
\usage{
convert_scan_to_df(scan, method = c("theoretical", "group", "focal"))
}
\arguments{
\item{scan}{a scan object}

\item{method}{character scalar, the scan.list to transform into an edge-list. Either "theoretical","group", or "focal"}
}
\value{
a data frame with columns i and j representing the association between node i and j (more or less considered always directed at present)
}
\description{
Convert a scan object into a data frame
Helper function to ensure compatibility with the ANTs package
}
\examples{
set.seed(42)

n<- 5;nodes<- letters[1:n];total_scan<- 42;
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[non.diagonal(Adj)]<- sample(0:total_scan,n*(n-1),replace = TRUE)
Adj

# by default will simulate a single directed theoretical scan
simu <- simu_scan(Adj,total_scan,"all")
convert_scan_to_df(simu)
}
