% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_conversion.R
\name{convert_to_igraph}
\alias{convert_to_igraph}
\title{Convert scanList 3D array into \code{igraph} object}
\usage{
convert_to_igraph(scan.list)
}
\arguments{
\item{scan.list}{a 3 dimensional array representing adjacency matrices (first 2 dimensions)
throughout the different scans (3rd dimension)}
}
\value{
an \code{igraph} object
}
\description{
Convert scanList 3D array into \code{igraph} object
}
\keyword{internal}
