% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_tools.R
\name{copy_attrs_to}
\alias{copy_attrs_to}
\title{\code{scanList}'s \code{attrs} attributes related convenience functions: copy attrs from one \code{scanList} to
another}
\usage{
copy_attrs_to(from, to)
}
\arguments{
\item{from}{a \code{scanList} object which \code{attrs} attribute to copy (see \code{\link[=simunet]{simunet()}})}

\item{to}{a \code{scanList} object to which \code{attrs} attribute should be pasted (see
\code{\link[=simunet]{simunet()}})}
}
\value{
a \code{scanList} object, the 3D array containted in \code{to} with \code{from}'s \code{attrs}
}
\description{
\code{scanList}'s \code{attrs} attributes related convenience functions: copy attrs from one \code{scanList} to
another
}
\keyword{internal}
