% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{count_nonNA}
\alias{count_nonNA}
\title{Count observed edges (non-\code{NA}s) for each edge in list of scans
Internal use. Used to determine the sampling effort across all scans performed}
\usage{
count_nonNA(scan.list, ...)
}
\arguments{
\item{scan.list}{a list of binary adjacency matrices, where an unobserved dyad (whether it is theoretically 0 or 1) is \code{NA}}

\item{...}{TO WRITE}
}
\value{
an integer matrix representing the sampling effort for each dyad
}
\description{
Count observed edges (non-\code{NA}s) for each edge in list of scans
Internal use. Used to determine the sampling effort across all scans performed
}
\examples{
# TO WRITE
}
