% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{customize_sampling}
\alias{customize_sampling}
\title{TO WRITE}
\usage{
customize_sampling(method = c("group", "focal"), obsProb)
}
\arguments{
\item{method}{TO WRITE}

\item{obsProb}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
