% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare.event.optimization_tools.R
\name{decide_use.rare.opti}
\alias{decide_use.rare.opti}
\title{Decide based on expected times if the otpimization for rare event should be used}
\usage{
decide_use.rare.opti(n, total_scan, max.obs = NULL, alpha = 0.05)
}
\arguments{
\item{n}{number of node of the original network, or can be an adjacency matrix}

\item{total_scan}{integer, sampling effort.}

\item{max.obs}{maximum value of the weighted adjacency matrix of the original network. Optional if n is inputted as an adjacency matrix.}

\item{alpha}{numerical in [0,1], type I error significance level.}
}
\value{
a logical value meaning that the optimization for rare events should be use when TRUE is returned.
}
\description{
Decide based on expected times if the otpimization for rare event should be used
}
\examples{
decide_use.rare.opti(30,900,10)
decide_use.rare.opti(30,9000,10)
}
