% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_mode_tools.R
\name{determine_Adj.subfun}
\alias{determine_Adj.subfun}
\title{Determine the right adjacency matrix subsetting function according to the chosen igraph mode}
\usage{
determine_Adj.subfun(mode)
}
\arguments{
\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. Default
here is directed. Possible values are: directed, undirected, upper, lower, max, min, plus.
Added vector too. See details
\code{\link[igraph:graph_from_adjacency_matrix]{graph_from_adjacency_matrix()}}.}
}
\value{
a subsetting function among \code{non.diagonal}, \code{upper.tri}, \code{lower.tri} or \code{function(V) {rep(TRUE,length(V))}}
}
\description{
Determine the right adjacency matrix subsetting function according to the chosen igraph mode
}
\keyword{internal}
