% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_sampling.R
\name{determine_sampling_type}
\alias{determine_sampling_type}
\title{Checks if the method and sampling parameter combination is adequate}
\usage{
determine_sampling_type(
  method,
  sampling = c("constant", "matrix", "even", "random", "function")
)
}
\arguments{
\item{method}{inputted \code{method} parameter, see \code{\link[=customize_sampling]{customize_sampling()}}}

\item{sampling}{inputted \code{sampling} parameter, see \code{\link[=customize_sampling]{customize_sampling()}}}
}
\value{
character, the type of sampling. Returns an error if incompatible method/sampling
combination
}
\description{
Checks if the method and sampling parameter combination is adequate
}
\keyword{internal}
