% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_tools.R
\name{draw_raw_scanList}
\alias{draw_raw_scanList}
\title{Draw edge presence according to the edge presence probability matrix}
\usage{
draw_raw_scanList(edge.Prob, n.scans)
}
\arguments{
\item{edge.Prob}{an \code{edgeProb} object, i.e. a list containing:
\itemize{
\item \code{P}: the edge presence probability matrix
\item \code{Adj}: the inputted \code{Adj}
\item \code{samp.effort}: the inputted \code{samp.effort}
\item \code{mode}: the inputted \code{mode}
\item \code{Adj.subfun}: the inputted \code{Adj.subfun}
}}

\item{n.scans}{integer scalar, number of scans to generate in the simulation}
}
\value{
a 3 dimensional array
representing the (binary) adjacency matrices (coded within the first two dimensions of the
3D-array) obtained at each simulated scan (coded as the 3rd dimension of the 3D-array)
}
\description{
Draw edge presence according to the edge presence probability matrix
}
\seealso{
\code{\link[=simunet]{simunet()}}, \code{\link[=generate_edgeProb]{generate_edgeProb()}}, \code{\link[=draw_edgeProb]{draw_edgeProb()}}, \code{\link[=generate_scanList]{generate_scanList()}}.
}
\keyword{internal}
