% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal.list.R
\name{focal.scan}
\alias{focal.scan}
\title{Focal scan from theoretical}
\usage{
focal.scan(scan.theoretical, focal)
}
\arguments{
\item{scan.theoretical}{binary matrix (the theoretical scan within do.scan())}

\item{focal}{either the row index or the name of the focal}
}
\value{
a focal scan as a binary matrix with matching values for the row of the focal, and NAs otherwise, and the name of the focal as a "focal" attribute.
}
\description{
Focal scan from theoretical
}
\examples{
# Internal use in do.(non.zero.)scan()
}
