% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_sampling.R
\name{focal_sample}
\alias{focal_sample}
\title{TO WRITE}
\usage{
focal_sample(
  scan.list,
  sampling = c("even", "random", "function"),
  all.sampled = TRUE
)
}
\arguments{
\item{scan.list}{TO WRITE}

\item{sampling}{TO WRITE}

\item{all.sampled}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
