% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_tools.R
\name{generate_empiscanList}
\alias{generate_empiscanList}
\title{Generator for \emph{empirical} \code{scanList} objects
Internal use. The user should rather rely on \code{\link[=simunet]{simunet()}} and/or
\code{\link[=perform_exp]{perform_exp()}} as a wrapper for the different steps needed to perform
simulations and experimental manipulations.}
\usage{
generate_empiscanList(scan.list, exp.design)
}
\arguments{
\item{scan.list}{a \code{scanList} object (see \code{\link[=simunet]{simunet()}})}

\item{exp.design}{an \code{expDesign} object. See objects returned by \code{\link[=design_exp]{design_exp()}}}
}
\value{
an \code{empirical} inheriting from \code{scanList} object, primarily a 3 dimensional array
representing the (binary) adjacency matrices (coded within the first two dimensions of the
3D-array) obtained at each simulated scan (coded as the 3rd dimension of the 3D-array), and a
list of attributes, \code{attrs}.

The list of attributes \code{attrs} contains:
\itemize{
\item all the previous attributes contained in \code{scan.list}'s \code{attrs} attributes list, as well as:
\itemize{
\item \code{scanList.type}: character scalar, changed from \code{"theoretical"} to \code{"empirical"}
\item \code{theoretical.scanList}: the 3D array \emph{before} the experimental manipulations contained in
the inputted \code{exp.design} have been applied
}
}
}
\description{
Generator for \emph{empirical} \code{scanList} objects
Internal use. The user should rather rely on \code{\link[=simunet]{simunet()}} and/or
\code{\link[=perform_exp]{perform_exp()}} as a wrapper for the different steps needed to perform
simulations and experimental manipulations.
}
\seealso{
\code{\link[=simunet]{simunet()}}, \code{\link[=generate_edgeProb]{generate_edgeProb()}}, \code{\link[=draw_edgeProb]{draw_edgeProb()}}, \code{\link[=generate_scanList]{generate_scanList()}}.
}
\keyword{internal}
