% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_tools.R
\name{generate_expDesign}
\alias{generate_expDesign}
\title{TO WRITE}
\usage{
generate_expDesign(FUN.seq)
}
\arguments{
\item{FUN.seq}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
