% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal.R
\name{generate_focal}
\alias{generate_focal}
\title{Generator for `focal` objects}
\usage{
generate_focal(focal.list, scan.number)
}
\arguments{
\item{focal.list}{a `focalList` object}

\item{scan.number}{an integer between in `1:total_scan`}
}
\value{
an `focal` object (S3 class) containing:
\itemize{
  \item{focal}{named integer scalar representing the _index_ of the node to sample at scan number `scan.number`}
  \item{focal.list}{inputted `focal.list` object}
  \item{scan.number}{inputted `scan.number`}
}
}
\description{
Generator for `focal` objects
}
\examples{

set.seed(42)

n<- 5;nodes<- letters[1:n];total_scan<- 42;
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[non.diagonal(Adj)]<- sample(0:total_scan,n*(n-1),replace = TRUE)
Adj

focal.list<- generate_focal.list(Adj,total_scan,focal.prob_fun = "even")
generate_focal(focal.list,5)
# below returns an error if `scan.number` is incompatible with `total_scan`
# generate_focal(focal.list,50)

}
