% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presenceProb.R
\name{generate_presenceProb}
\alias{generate_presenceProb}
\title{Generator for `presenceProb` objects}
\usage{
generate_presenceProb(Adj, total_scan, mode, Adj.subfun = NULL)
}
\arguments{
\item{Adj}{square integers matrix of occurrences of dyads.}

\item{total_scan}{integer, sampling effort. Note that 1/total_scan should be relatively small, increasingly small with increasing precision. Optional if using presence.prob.}

\item{mode}{Character scalar, specifies what type of igraph network `mode` should be used to convert the supplied matrix. Possible values are:
\itemize{
  \item{`"directed"` (the default): for non-symetrical adjacency matrix where `Adj[i,j]` doesn't have the same meaning as `Adj[j,i]`}
  \item{`"undirected"`: same as `"max"`}
  \item{`"upper"`: undirected matrix focusing only on the upper triangle of `Adj` (relying on `upper.tri`). Either `"upper"` or `"lower"` could be favor if only one of `Adj[i,j]` and `Adj[j,i]` should be randomized}
  \item{`"lower"`: undirected matrix focusing only on the lower triangle of `Adj` (relying on `lower.tri`)}
  \item{`"max"`: from a `"directed"` randomization process (both `Adj[i,j]` and `Adj[j,i]` will be drawn at each scan), `max(Adj[i,j],Adj[j,i])` will be kept for both}
  \item{`"min"`: from a `"directed"` randomization process (both `Adj[i,j]` and `Adj[j,i]` will be drawn at each scan), `min(Adj[i,j],Adj[j,i])` will be kept for both}
  \item{`"plus."`:  from a `"directed"` randomization process (both `Adj[i,j]` and `Adj[j,i]` will be drawn at each scan), `Adj[i,j] + Adj[j,i]` will be kept for both}
  \item{`"vector"`: experimental. To consider adjacency matrices as flat vectors to be randomized. Relevance unexplored yet.}
  \item{See details \link[igraph]{graph_from_adjacency_matrix}}
}}

\item{Adj.subfun}{optional, used internally in scan-related functions. Subsetting function of the adjacency matrix. Driven by igraph `"mode"` argument.}
}
\value{
a `presenceProb` object (S3 class) containing:
\itemize{
  \item{`P`: a [0,1] numeric matrix of probability of presence of a tie between each dyad}
  \item{`Adj`: inputted `Adj`}
  \item{`total_scan`: inputted `total_scan`}
  \item{`mode`: inputted `mode`}
  \item{`Adj.subfun`: inputted `Adj.subfun`}
}
}
\description{
Generator for `presenceProb` objects
}
\examples{
set.seed(42)

n<- 5;nodes<- as.character(1:n);total_scan<- 42;
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[non.diagonal(Adj)]<- sample(0:total_scan,n*(n-1),replace = TRUE)
Adj

generate_presenceProb(Adj,total_scan,mode = "directed")
}
