% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_tools.R
\name{generate_scanList}
\alias{generate_scanList}
\title{Generator for \code{scanList} objects
Internal use. The user should rather rely on \code{simunet} as a wrapper for the different steps
needed to perform the simulations from inputted data.}
\usage{
generate_scanList(edge.Prob, n.scans)
}
\arguments{
\item{edge.Prob}{an \code{edgeProb} object, i.e. a list containing:
\itemize{
\item \code{P}: the edge presence probability matrix
\item \code{Adj}: the inputted \code{Adj}
\item \code{samp.effort}: the inputted \code{samp.effort}
\item \code{mode}: the inputted \code{mode}
\item \code{Adj.subfun}: the inputted \code{Adj.subfun}
}}

\item{samp.effort}{integer scalar, the sampling effort, or number of scans, that led to obtaining
of \code{Adj}}
}
\value{
a \code{theoretical} inheriting from \code{scanList} object, primarily a 3 dimensional array
representing the (binary) adjacency matrices (coded within the first two dimensions of the
3D-array) obtained at each simulated scan (coded as the 3rd dimension of the 3D-array), and a
list of attributes, \code{attrs}.

The list of attributes \code{attrs} contains:
\itemize{
\item \code{scanList.type}: character scalar, \code{"theoretical"} at first and \code{"empirical"} after a non-\code{NULL}
experimental manipulation has been applied to the \code{scanList} (via \code{\link[=perform_exp]{perform_exp()}}
and a \code{expDesign} object)
\item \code{raw.scanList}: the 3D binary array, directed, before potential symmetrization attempt by
applying the igraph's mode via \code{\link[=apply_mode]{apply_mode()}}
\item \code{Adj}: integer matrix, \code{Adj} contained in \code{edge.Prob}
\item \code{samp.effort}: integer, \code{samp.effort} contained in \code{edge.Prob}
\item \code{n.scans}: inputted \code{n.scans}
\item \code{mode}: character scalar, \code{mode} contained in \code{edge.Prob}
\item \code{Adj.subfun}: function, \code{Adj.subfun} contained in \code{edge.Prob}
\item \code{edge.Prob}: numeric matrix,\code{edge.Prob$P} (only the probability matrix) data contained in
\code{edge.Prob}
}
}
\description{
Generator for \code{scanList} objects
Internal use. The user should rather rely on \code{simunet} as a wrapper for the different steps
needed to perform the simulations from inputted data.
}
\seealso{
\code{\link[=simunet]{simunet()}}, \code{\link[=generate_edgeProb]{generate_edgeProb()}}, \code{\link[=draw_edgeProb]{draw_edgeProb()}}, \code{\link[=generate_empiscanList]{generate_empiscanList()}}.
}
\keyword{internal}
