% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_tools.R
\name{get_attrs}
\alias{get_attrs}
\title{\code{scanList}'s \code{attrs} attributes related convenience functions: retrieve all attributes}
\usage{
get_attrs(scan.list)
}
\arguments{
\item{scan.list}{a \code{scanList} object (see \code{\link[=simunet]{simunet()}})}
}
\value{
list, the list of attributes stored in \code{scan.list}'s \code{attrs} attribute
}
\description{
\code{scanList}'s \code{attrs} attributes related convenience functions: retrieve all attributes
}
\keyword{internal}
