% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{group_samp.list}
\alias{group_samp.list}
\title{TO WRITE}
\usage{
group_samp.list(scan.list, obs.P)
}
\arguments{
\item{scan.list}{TO WRITE}

\item{obs.P}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
