% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_matList_conversion.R
\name{igraph2scanList}
\alias{igraph2scanList}
\title{Convert \code{igraphSN} into \code{scanList} objects}
\usage{
igraph2scanList(G, format = c("weightedAdj", "scanList"))
}
\arguments{
\item{G}{an \code{igraphSN} object (see \code{\link[=scanList2igraph]{scanList2igraph()}})}

\item{format}{character, option to specify if a \code{weightedAdj} or a 3D binary \code{scanList} should be
returned}
}
\value{
scanList object. See \code{\link[=simunet]{simunet()}}
}
\description{
Convenience function to back-transform a \code{igraphSN} object (\code{igraph} object with \code{attrs}) into a
\code{scanList} object
}
\examples{
set.seed(42)
n <- 5L
samp.effort <- 100L

# Adjacency matrix import
## random directed adjacency matrix
Adj <- sample(1:samp.effort,n * n) |>
  matrix(nrow = 5,dimnames = list(letters[1:n],letters[1:n]))
Adj[lower.tri(Adj,diag = TRUE)] <- 0L
Adj

sL <- simunet(Adj,samp.effort,"upper",10)
G <- sL |> scanList2igraph()
G
G |> igraph2scanList()
G |> igraph2scanList(format = "scanList")
identical(G |> igraph2scanList(format = "scanList"),sL)
}
