% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asnr_tools.R
\name{import_from_asnr}
\alias{import_from_asnr}
\title{Import network from asnr package
Bansal lab's Animal Social Network Repository}
\usage{
import_from_asnr(
  class = NULL,
  species = NULL,
  network = NULL,
  url = NULL,
  output = c("graph", "adjacency"),
  type = c("both", "upper", "lower"),
  ...,
  default_prefix = "https://raw.githubusercontent.com/bansallab/asnr/master/Networks/",
  full.path = NULL
)
}
\arguments{
\item{class}{TO WRITE}

\item{species}{TO WRITE}

\item{network}{TO WRITE}

\item{url}{TO WRITE}

\item{output}{TO WRITE}

\item{type}{TO WRITE}

\item{...}{TO WRITE}

\item{default_prefix}{TO WRITE}

\item{full.path}{TO WRITE}
}
\value{
TO WRITE
}
\description{
Import network from asnr package
Bansal lab's Animal Social Network Repository
}
\examples{
import_from_asnr("Aves","cowbird")
# partial matching works, but at the moment not case insensitive
import_from_asnr("Amph","frog",output = "adj")
# Users can also copy-paste a `.graphml`'s URL
import_from_asnr(url = "https://github.com/bansallab/asnr/blob/master/Networks/Reptilia/
lizard_proximity_weighted/weighted_network_social_T_rugosa.graphml")
}
