% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.Obs.list.R
\name{make.Obs.list}
\alias{make.Obs.list}
\title{Make the list of Observation subsets}
\usage{
make.Obs.list(DT, Obs, ID.present = TRUE, n.core = parallel::detectCores() - 1)
}
\arguments{
\item{DT}{data.table of the events/samples of interest, for each and relative to which subsets of data will be produced. Subset between datesdinf and dsup.}

\item{Obs}{data.table of dyadic observations with their date, and the two nodes in column id and tar (for focal id and target of interaction)}

\item{ID.present}{logical. If TRUE, provide the subset for a given event/sample only if the individual of intrest from DT is present at least once in a subset of dyadic obseravtions.}

\item{n.core}{number of (logical) core to use. Default is all minus one.}
}
\value{
a list of subset dyadic observations data table. If no observation were within [dinf,sdup], returns an NA with the reason as an attribute. If data were present but ID.present is TRUE, return an empty data.table with the reason as an attribute.
}
\description{
Subset the dyadic observations into event/sample-centric observations datasets
}
\examples{
set.seed(42);ID<- letters[1:10];

D<- round(runif(n = 200,min = 1,max = 30));
DT<- data.table::data.table(date=as.Date(D,origin = "1991-03-30"),
                            id=sample(ID,200,replace = TRUE));
DT<- add.prepatent(DT,mean.lag = 0, duration = 15,extension = 0,unit = "day");

D.Obs<- round(runif(n = 4000,min = 0,max = 15))
Obs<- data.table::data.table(date=as.Date(D.Obs-15,origin = "1991-03-30"),
                             id=sample(ID,1000,replace = TRUE),
                             tar=sample(ID,1000,replace = TRUE))

make.Obs.list(DT,Obs)
# multi.core<- system.time(make.Obs.list(DT,Obs,ID.present = FALSE))
# multi.core
# single.core<- system.time(make.Obs.list(DT,Obs,ID.present = FALSE,n.core = 1))
# single.core
}
