% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{make_cl}
\alias{make_cl}
\title{Make cluster object wrapper}
\usage{
make_cl(n.cores, .export)
}
\arguments{
\item{n.cores}{number of threads to use.}

\item{.export}{vector of variable/function names to use in each clusters}
}
\value{
a cluster object
}
\description{
Make cluster object wrapper
}
\examples{
# Internal use
}
