% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_sampling.R
\name{mask_non.focals}
\alias{mask_non.focals}
\title{Mask edges not-involving the scan's focal node (applies the focal list)
Internal.}
\usage{
mask_non.focals(scan.list, focalList)
}
\arguments{
\item{scan.list}{a \code{scanList} object}

\item{focalList}{named integer vector, returned by draw_focalList representing the node's index
(row/column) to be sampled for each scan. Names are obtain from the adjacency matrix \code{Adj}, the
vector's length is equal to \code{n.scans}}
}
\value{
a 3D array where edges not-involving the scan's focal node are replaced by\code{NA}s
}
\description{
Mask edges not-involving the scan's focal node (applies the focal list)
Internal.
}
\keyword{internal}
