% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_matList_conversion.R
\name{matList2array}
\alias{matList2array}
\title{Convert list of 2D matrices into scanList 3D array}
\usage{
matList2array(mat.list)
}
\arguments{
\item{mat.list}{a 3 dimensional array representing adjacency matrices (first 2 dimensions)
throughout the different scans (3rd dimension)}
}
\value{
a 3 dimensional array representing adjacency matrices (first 2 dimensions) throughout the
different scans (3rd dimension)
}
\description{
Convenience function to back-transform a list of matrices into a 3D array
}
\keyword{internal}
