% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_matList_conversion.R
\name{matList2scanList}
\alias{matList2scanList}
\title{Convert \code{matList} into \code{scanList} objects}
\usage{
matList2scanList(mat.list)
}
\arguments{
\item{mat.list}{a 3 dimensional array representing adjacency matrices (first 2 dimensions)
throughout the different scans (3rd dimension)}
}
\value{
scanList object. See \code{\link[=simunet]{simunet()}}
}
\description{
Convenience function to back-transform a \code{matList} object (list of matrices) into a \code{scanList}
object
}
\examples{
set.seed(42)
n <- 5L
samp.effort <- 100L

# Adjacency matrix import
## random directed adjacency matrix
Adj <- sample(1:samp.effort,n * n) |>
  matrix(nrow = 5,dimnames = list(letters[1:n],letters[1:n]))
Adj[lower.tri(Adj,diag = TRUE)] <- 0L
Adj

sL <- simunet(Adj,samp.effort,"upper",10)
mL <- sL |> scanList2matList()
mL
mL |> matList2scanList()
identical(mL |> matList2scanList(),sL)
}
