% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.tools.R
\name{matrix_sum_na.rm}
\alias{matrix_sum_na.rm}
\title{Matrix sum removing NAs
Equivalent to the elmement-wise matrix addition, but replacing NAs by zeros. Internal use}
\usage{
matrix_sum_na.rm(X, Y)
}
\arguments{
\item{X}{first matrix}

\item{Y}{second matrix, similarly dimensioned}
}
\value{
the sum of X and Y, replacing NAs by zeros
}
\description{
Matrix sum removing NAs
Equivalent to the elmement-wise matrix addition, but replacing NAs by zeros. Internal use
}
\examples{
#Internal use
}
