% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_up.scans.R
\name{n.observed_edges}
\alias{n.observed_edges}
\title{Get number of edge observations (for group scans with unobservable individuals)
quantify actual edge-wise sampling effort, considering that some weren't observable in all group scans.Internal use.}
\usage{
n.observed_edges(
  scan_list,
  mode,
  diag = NULL,
  use.rare.opti = FALSE,
  obs.prob = NULL,
  focal.list = NULL,
  n.zeros = 0,
  method = NULL
)
}
\arguments{
\item{scan_list}{list of binary group scans, with NAs when the dyad was not observable.}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. Default here is directed. Possible values are: directed, undirected, upper, lower, max, min, plus. Added vector too. See details \link[igraph]{graph_from_adjacency_matrix}.}

\item{diag}{integer (mostly), value to replace the diagonal of the output matrix with. Use NULL if you consider self-loop (untested).}

\item{use.rare.opti}{logical: should the optimization for rare event be used?}

\item{obs.prob}{either :
\itemize{
 \item{"a dyad observation probability matrix (P.obs)"}{of same dimension as Adj}
 \item{"a dyad observation vector"}{subsetted similarly as Adj (through the non.diagonal() function for instance)}
 \item{"a systematic dyad observation (P.obs constant for all i,j)"}{should be in [0,1], assumed to be the case when only one value is inputed)}
}}

\item{focal.list}{Character vector, indicate the list of focals to consider throughout the scans.}

\item{n.zeros}{integer, the attribute outputed by `simulate_zeros.non.zeros`, representing the number of full-zero scans. Used only when use.rare.opti=TRUE}

\item{method}{Character scalar, specify if the function should use a whole group or a focal scan sampling method (or both).}
}
\value{
a square matrix with element quantifying how many time a dyad has been sampled
}
\description{
Get number of edge observations (for group scans with unobservable individuals)
quantify actual edge-wise sampling effort, considering that some weren't observable in all group scans.Internal use.
}
\examples{
#internal use.
}
