% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_tools.R
\name{namedList}
\alias{namedList}
\title{Helper to manage function names in list of functions
from \url{https://stackoverflow.com/a/16951524}}
\usage{
namedList(...)
}
\arguments{
\item{...}{functions of \code{scanList} objects. Can be a combination of the provided "building
blocks" and user-defined function (cf. the example section).}
}
\value{
a named list of functions
}
\description{
Helper to manage function names in list of functions
from \url{https://stackoverflow.com/a/16951524}
}
\keyword{internal}
