% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare.event.optimization_tools.R
\name{new.data.poly}
\alias{new.data.poly}
\title{Generate new data using orthogonal polynoms of input variable}
\usage{
new.data.poly(
  n,
  total_scan,
  max.obs,
  algorithm = c("standard", "optimization for rare event")
)
}
\arguments{
\item{n}{integer, number of node of the network}

\item{total_scan}{integer, sampling effort}

\item{max.obs}{integer, maximum number of observation after `total_scan` scans}

\item{algorithm}{character, which algorithm to produce new data for.}
}
\value{
a data.frame inputed on the same orthogonal polynomial scale as the one used by the standard and optimized models of expected time
}
\description{
Generate new data using orthogonal polynoms of input variable
}
\examples{
# Internal use in opti.expected.time
}
