% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{non.diagonal<-}
\alias{non.diagonal<-}
\title{TO WRITE}
\usage{
non.diagonal(x) <- value
}
\arguments{
\item{x}{matrix or other R object with `length(dim(x)) == 2`}

\item{value}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
M <- matrix(sample(1:10,16,replace = TRUE),4,4)
M
non.diagonal(M) <- runif(12,0,1)
M
non.diagonal(M) <- 42
M
}
