% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{non.diagonal}
\alias{non.diagonal}
\title{Non Diagonal Part of a Matrix}
\usage{
non.diagonal(M, output = c("matrix.logical", "vector.values"))
}
\arguments{
\item{M}{square matrix.}

\item{output}{choose to either return a matrix of logical values (\code{TRUE}s, \code{FALSE}s on the diagonal, i.e. comparable behavior than \code{upper.tri} for instance) or a vector of the subsetted values of the matrix.}
}
\value{
either return a matrix of logical values (\code{TRUE}s, \code{FALSE}s on the diagonal) or a vector of the subsetted values of the matrix.
}
\description{
Similarly to \code{upper.tri} and \code{lower.tri}, returns a matrix of \code{logical}s to identify the diagonal of a square matrix
}
\examples{
M<- matrix(sample(1:10,16,replace = TRUE),4,4)
non.diagonal(M)
}
