% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.tools.R
\name{non.zero.non.diag}
\alias{non.zero.non.diag}
\title{Identify coordinates of non null non diagonal elements of a matrix}
\usage{
non.zero.non.diag(M)
}
\arguments{
\item{M}{a square matrix}
}
\value{
an array of row and col indices of the positive and non diagonal elements
}
\description{
Identify coordinates of non null non diagonal elements of a matrix
}
\examples{
set.seed(42)
M<- matrix(sample(0:10,16,replace = TRUE),4,4)

non.zero.non.diag(M)

}
