% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.scan.R
\name{null_na}
\alias{null_na}
\title{Replace NAs by zeros in vectors/matrices}
\usage{
null_na(X)
}
\arguments{
\item{X}{a vector or matrix}
}
\value{
similarly dimensioned vector or matrix with zeros instead of NAs
}
\description{
Replace NAs by zeros in vectors/matrices
}
\examples{
# Internal use
}
