% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.prob_tools.R
\name{observable_edges}
\alias{observable_edges}
\title{Hide unobservable edges
Simulate that some dyads might reasonable not be observable during a group scan}
\usage{
observable_edges(Scan, obs.prob = NULL, Adj.subfun = NULL)
}
\arguments{
\item{Scan}{a square matrix}

\item{obs.prob}{either :
\itemize{
 \item{"a dyad observation obs.probability matrix"}{of same dimension as Scan}
 \item{"a dyad observation vector"}{subsetted similarly as Scan (through the non.diagonal() function for instance)}
 \item{"a systematic dyad observation obs.probability"}{should be in [0,1], assumed to be the case when only one value is inputed)}
}}

\item{Adj.subfun}{subsetting function of the adjacency matrix. Driven by igraph "mode" argument}
}
\value{
a similar matrix as Scan where some non diagonal edges have a obs.probability to be NAs.
}
\description{
Hide unobservable edges
Simulate that some dyads might reasonable not be observable during a group scan
}
\examples{
set.seed(42)

n<- 6;nodes<- as.character(1:n);
Scan<- matrix(rbinom(n*n,1,0.2),n,n,dimnames = list(nodes,nodes));diag(Scan)<- 0;
obs.prob<- matrix(runif(n*n,0,1),n,n);diag(obs.prob)<- 0

observable_edges(Scan,obs.prob,non.diagonal)
}
