% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_tools.R
\name{perform_exp}
\alias{perform_exp}
\title{TO WRITE}
\usage{
perform_exp(scan.list, exp.design = NULL)
}
\arguments{
\item{scan.list}{TO WRITE}

\item{exp.design}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
