% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{proportional.prob}
\alias{proportional.prob}
\title{Modify a vector to be probabilities in ]0,1[}
\usage{
proportional.prob(P)
}
\arguments{
\item{P}{a vector of values not all in ]0,1[}
}
\value{
a vector of valid probabilities in ]0,1[
}
\description{
Modify a vector to be probabilities in ]0,1[
}
\examples{
# internal use in make_obs.prob().
}
