% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{quick.sample}
\alias{quick.sample}
\title{Quick optimized equivalent to sample(x,size,replace=TRUE)}
\usage{
quick.sample(x, size)
}
\arguments{
\item{x}{a vector}

\item{size}{number of elements to sample}
}
\value{
sample of size "size" taken from x
}
\description{
Quick optimized equivalent to sample(x,size,replace=TRUE)
}
\examples{
quick.sample(1:20,5)
# microbenchmark::microbenchmark(runif={(1:20)[ceiling(runif(5,0,20))]},
#   quick.sample=quick.sample(1:20,5),sample=sample(1:20,5,replace = TRUE),
#   times = 1000,control = list("warmup"=100))
}
