% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{quick_sample}
\alias{quick_sample}
\title{Quick optimized equivalent to sample(x,size,replace=TRUE)}
\usage{
quick_sample(x, size)
}
\arguments{
\item{x}{a vector}

\item{size}{number of elements to sample}
}
\value{
sample of size "size" taken from x
}
\description{
Quick optimized equivalent to sample(x,size,replace=TRUE)
}
\examples{
quick_sample(1:20,5)
}
