% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{rbind_lapply}
\alias{rbind_lapply}
\title{Row bind list of data frames
wrapper to one-function do.call rbind over a lapply list}
\usage{
rbind_lapply(X, FUN)
}
\arguments{
\item{X}{a list. See details \link[base]{lapply}.}

\item{FUN}{a function to subset data frames (or data tables). See details \link[base]{lapply}.}
}
\value{
a row bound data frame
}
\description{
Row bind list of data frames
wrapper to one-function do.call rbind over a lapply list
}
\examples{
set.seed(42)

X<- lapply(1:3,function(i) list(int = 42,df = data.frame(x = runif(10,0,1),y = runif(10,0,1))))
rbind_lapply(X,function(x) x$df)
}
