% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{rbind_pblapply}
\alias{rbind_pblapply}
\title{Row bind list of data frames
wrapper to one-function do.call rbind over a pblapply list}
\usage{
rbind_pblapply(X, FUN, n.cores = NULL, .export = NULL, cl = NULL)
}
\arguments{
\item{X}{a list. See details \link[base]{lapply}.}

\item{FUN}{a function to subset data frames (or data tables). See details \link[base]{lapply}.}

\item{n.cores}{number or cores (threads rather) to use in the cluster.}

\item{.export}{character vector of the name of the variable and function to export to each worker of the cluster}

\item{cl}{cluster object}
}
\value{
a row bound data frame
}
\description{
Row bind list of data frames
wrapper to one-function do.call rbind over a pblapply list
}
\examples{
set.seed(42)

X<- lapply(1:3,function(i) list(int = 42,df = data.frame(x = runif(10,0,1),y = runif(10,0,1))))
rbind_lapply(X,function(x) x$df)
}
