% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{remove_mostPeripheral}
\alias{remove_mostPeripheral}
\title{TO WRITE}
\usage{
remove_mostPeripheral(scan.list, ...)
}
\arguments{
\item{scan.list}{TO WRITE}

\item{...}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
