% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{remove_mostPeripheral.sLlist}
\alias{remove_mostPeripheral.sLlist}
\title{TO WRITE}
\usage{
\method{remove_mostPeripheral}{sLlist}(scan.list)
}
\arguments{
\item{scan.list}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
