% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_NA_tools.R
\name{resolve_NA}
\alias{resolve_NA}
\title{Replace \code{NA}s with known values when they can be determined
Internal use. When the \code{"min"} or \code{"max"} algorithm is chosen to convert directed networks into
undirected one, some \code{NA}s can be deduced from their transposed values (cf. below).}
\usage{
resolve_NA(
  scan.list,
  mode = c("directed", "undirected", "max", "min", "upper", "lower", "plus", "vector")
)
}
\arguments{
\item{scan.list}{a \code{scanList} object. See objects returned by \code{\link[=simunet]{simunet()}}}

\item{mode}{character scalar, See \code{\link[=simunet]{simunet()}}}
}
\value{
a list of binary adjacency matrix where resolvable \code{NA}s are changed depending on the
chosen mode:
\itemize{
\item \verb{scan[i,j] = NA & scan[j,i] = 0 => scan[i,j] = 0} when \code{mode = "min"}
\item \verb{scan[i,j] = NA & scan[j,i] = 1 => scan[i,j] = 1} when \code{mode = "max"}
}
}
\description{
Replace \code{NA}s with known values when they can be determined
Internal use. When the \code{"min"} or \code{"max"} algorithm is chosen to convert directed networks into
undirected one, some \code{NA}s can be deduced from their transposed values (cf. below).
}
\keyword{internal}
