% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{scale_scans}
\alias{scale_scans}
\title{TO WRITE}
\usage{
scale_scans(summed, ...)
}
\arguments{
\item{summed}{TO WRITE}

\item{...}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
