% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{scale_scans.sLlist}
\alias{scale_scans.sLlist}
\title{TO WRITE}
\usage{
\method{scale_scans}{sLlist}(summed)
}
\arguments{
\item{summed}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
