% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.scan.R
\name{scan.default.args}
\alias{scan.default.args}
\title{Set arguments to default for do.(non.zero.)scan() function when necessary}
\usage{
scan.default.args(Adj, total_scan, method, ...)
}
\arguments{
\item{Adj}{square integers matrix of occurences of dyads. Optional if using presence.prob. WIP: implement method for association matrices...}

\item{total_scan}{integer, sampling effort. Note that 1/total_scan should be relatively small, increasingly small with increasing precision. Optional if using presence.prob.}

\item{method}{Character scalar, specifies if the function should return a theoretical perfect group scan, an  empirical group scan (a similarly dimensioned matrix as Adj), or a focal scan (a vector representing the given focal's row in the group scan matrix).}

\item{...}{additional argument to be used, to use produce a scan in a desired way.#'
\itemize{
  \item{obs.prob}{either :
    \item{"a dyad observation obs.probability matrix"}{of same dimension as Adj}
    \item{"a dyad observation vector"}{subsetted similarly as Adj (through the non.diagonal() function for instance)}
    \item{"a general dyad observation obs.probability"}{should be in [0,1], assumed to be the case when only one value is inputed)}
  }
  \item{focal}{Only required for method = "focal" or "both" Character scalar, indicate which focal to consider for the scan.}
  \item{Adj.subfun}{subsetting function of the adjacency matrix. Driven by igraph "mode" argument}
  \item{presence.prob} {square probability matrix of presence (as in Bernouilli trial) of dyads. Optional if using Adj and total_scan.}
  \item{mode} {Character scalar, specifies how igraph should interpret the supplied matrix. Default here is directed. Possible values are: directed, undirected, upper, lower, max, min, plus. Added vector too. See details \link[igraph]{graph_from_adjacency_matrix}.}
}}
}
\value{
nothing, but assign required variable for do.(non.zero.)scan() in their environment.
}
\description{
Set arguments to default for do.(non.zero.)scan() function when necessary
}
\examples{
# Internal use.
}
