% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_matList_conversion.R
\name{scanList2matList}
\alias{scanList2matList}
\title{Convert scanList 3D array into list of 2D matrices}
\usage{
scanList2matList(scan.list)
}
\arguments{
\item{scan.list}{a 3 dimensional array representing adjacency matrices (first 2 dimensions)
throughout the different scans (3rd dimension)}
}
\value{
a list of matrices where each element is the 2D matrix for each scan index (3rd dimension of \code{array.3D})
}
\description{
Convenience function to allow the use of functions designed to affect list of matrices instead of
3D array (specifically \code{scanList} objects) where the 3rd dimension is the scan index. Keeps track
of the attributes list \code{attrs}
}
\details{
To back-transform a list of matrices into a \code{scanList} object, users can use
\code{matList2array()}, wrapper to \code{\link[=simplify2array]{simplify2array()}} that
handles attributes list \code{attrs}
}
\examples{
set.seed(42)
n <- 5L
samp.effort <- 100L

# Adjacency matrix import
## random directed adjacency matrix
Adj <- sample(1:samp.effort,n * n) |>
  matrix(nrow = 5,dimnames = list(letters[1:n],letters[1:n]))
Adj[lower.tri(Adj,diag = TRUE)] <- 0L
Adj
mL <- simunet(Adj,samp.effort,"upper",10) |> scanList2matList()
mL
mL |> matList2scanList()
}
