% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare.event.optimization_tools.R
\name{simulate_zeros.non.zeros}
\alias{simulate_zeros.non.zeros}
\title{Simulate which scan returns an all-zeros matrix}
\usage{
simulate_zeros.non.zeros(total_scan, presence.prob)
}
\arguments{
\item{total_scan}{integer, sampling effort}

\item{presence.prob}{presence probability matrix (or vector)}
}
\value{
a list of NULL representing the non-zero scan to run with an attribute `n.zero` being the number of full-zero scans,and `non.zero.pos` TO EDIT?
}
\description{
Simulate which scan returns an all-zeros matrix
}
\examples{
# Internal use
}
