% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{sum_scans.scanList}
\alias{sum_scans.scanList}
\title{TO WRITE}
\usage{
\method{sum_scans}{scanList}(scan.list, which = c("auto", "theoretical", "raw"))
}
\arguments{
\item{scan.list}{TO WRITE}

\item{which}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
