% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_tools.R
\name{t.scanList}
\alias{t.scanList}
\title{TO WRITE}
\usage{
\method{t}{scanList}(x)
}
\arguments{
\item{x}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
\keyword{internal}
