% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{t_test_from_summary}
\alias{t_test_from_summary}
\title{Two-sample t-tests from sample statistics
cf. https://stats.stackexchange.com/a/30450/255116}
\usage{
t_test_from_summary(m1, m2, sd1, sd2, n1, n2, m0 = 0, equal.variance = FALSE)
}
\arguments{
\item{m1}{numeric, sample mean in population 1.}

\item{m2}{numeric, sample mean in population 2.}

\item{sd1}{numeric, sample standard deviation in population 1.}

\item{sd2}{numeric, sample standard deviation in population 2.}

\item{n1}{integer, sample size of population 1.}

\item{n2}{integer, sample size of population 1.}

\item{m0}{numeric, null value for the difference in means to be tested for. Default is 0.}

\item{equal.variance}{logical, whether or not to assume equal variance. Default is FALSE.}
}
\value{
vector of `Difference of means`, `Std Error`, `t`, `p-value` of the t-test from summary data.
}
\description{
Two-sample t-tests from sample statistics
cf. https://stats.stackexchange.com/a/30450/255116
}
\examples{
t_test_from_summary(15,18,2,1.8,20,23)
}
