% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_conversion.R
\name{weightedAdj2scanList}
\alias{weightedAdj2scanList}
\title{Convert \code{weightedAdj} into 3D binary array \code{scanList} objects
Back-transform \code{weightedAdj}into the \code{scanList} it was summed from (see \code{\link[=sum_scans]{sum_scans()}})}
\usage{
weightedAdj2scanList(sum)
}
\arguments{
\item{sum}{a \code{weightedAdj} object}
}
\value{
\code{scanList} object. See \code{\link[=simunet]{simunet()}}
}
\description{
Convert \code{weightedAdj} into 3D binary array \code{scanList} objects
Back-transform \code{weightedAdj}into the \code{scanList} it was summed from (see \code{\link[=sum_scans]{sum_scans()}})
}
\examples{
set.seed(42)
n <- 5L
samp.effort <- 100L

# Adjacency matrix import
## random directed adjacency matrix
Adj <- sample(1:samp.effort,n * n) |>
  matrix(nrow = 5,dimnames = list(letters[1:n],letters[1:n]))
Adj[lower.tri(Adj,diag = TRUE)] <- 0L
Adj

sL <- simunet(Adj,samp.effort,"upper",10)
wAdj <- sL |> sum_scans()
wAdj
wAdj |> weightedAdj2scanList()
identical(wAdj |> weightedAdj2scanList(),sL)
}
