% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asnr_tools.R
\name{asnr_network_df}
\alias{asnr_network_df}
\title{Generate a data frame of all networks in the Animal Social Network Repository (asnr)}
\usage{
asnr_network_df(user = "bansallab", repo = "asnr")
}
\arguments{
\item{user}{character scalar, used to reconstruct the asnr github repository.
Default is "bansallab".}

\item{repo}{character scalar, used to reconstruct the asnr github repository.
Default is "asnr".}
}
\value{
a character data frame with the column "class", "species", and
"network"
}
\description{
Generate a data frame of all networks in the Animal Social Network Repository (asnr)
}
\examples{
asnr_network_df()
}
