% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_tools.R
\name{attrs}
\alias{attrs}
\title{\code{scanList}'s \code{attrs} attributes related convenience functions: retrieve or modify attributes
\code{attrs()} and \verb{attrs()<-} can be used to retrieve the named attributes contained in the
attributes list \code{attrs} of a \code{scanList} object}
\usage{
attrs(scan.list, a = NULL)
}
\arguments{
\item{scan.list}{a \code{scanList} object (see \code{\link[=simunet]{simunet()}})}

\item{a}{character (scalar or vector), the name(s) of the attribute(s) to retrieve, modify or add}
}
\value{
the attribute(s) requested, or the \code{scan.list} which \code{attrs} attribute has been modified
}
\description{
\code{scanList}'s \code{attrs} attributes related convenience functions: retrieve or modify attributes
\code{attrs()} and \verb{attrs()<-} can be used to retrieve the named attributes contained in the
attributes list \code{attrs} of a \code{scanList} object
}
\examples{
set.seed(42)
n <- 5L
samp.effort <- 100L

# Adjacency matrix import
## random directed adjacency matrix
Adj <- sample(1:samp.effort,n * n) |>
  matrix(nrow = 5,dimnames = list(letters[1:n],letters[1:n]))
Adj[lower.tri(Adj,diag = TRUE)] <- 0L
Adj

sL <- simunet(Adj = Adj,samp.effort = samp.effort,mode = "upper",n.scans = 120L)

# retrieve all attributes in `attrs`
sL |> attrs()

# retrieve a specific attribute from `attrs`
sL |> attrs("edge.Prob")

# modify a specific attribute from `attrs` (internal use)
attrs(sL,"scanList.type") <- "empirical"
attrs(sL,"scanList.type")
}
