% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_sampling.R
\name{check_sampling_parameters}
\alias{check_sampling_parameters}
\title{Checks if the method and sampling parameter combination is adequate}
\usage{
check_sampling_parameters(
  method,
  sampling = c("constant", "matrix", "even", "random", "function")
)
}
\arguments{
\item{method}{inputted \code{method} parameter, see \code{\link[=customize_sampling]{customize_sampling()}}}

\item{sampling}{inputted \code{sampling} parameter, see \code{\link[=customize_sampling]{customize_sampling()}}}
}
\value{
nothing, but returns an error if the combination is not adequate
}
\description{
Checks if the method and sampling parameter combination is adequate
}
\keyword{internal}
