% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{count_NA}
\alias{count_NA}
\title{TO WRITE}
\usage{
count_NA(scan.list, empirical = TRUE)
}
\arguments{
\item{scan.list}{TO WRITE}

\item{empirical}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
