% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{count_NA}
\alias{count_NA}
\title{Count number of times edges have not been observed
For empirical \code{scanList}, an edge that is not observed during a scan is \code{NA}. This function
counts how many time this was the case in the inputted \code{scan.list} for each edge.}
\usage{
count_NA(scan.list, ...)
}
\arguments{
\item{scan.list}{a \code{scanList} or \code{sLlist} object, where an unobserved edge (whether it is 0 or 1)
is \code{NA}. See objects returned by \code{\link[=simunet]{simunet()}}}

\item{...}{additional arguments to be passed. At the moment \code{scale_scans()} does not use

At the moment \code{count_NA()} does not use additional argument, arguments passed will be ignored.}
}
\value{
an integer matrix, or list of such, representing how many time each edge has been
unobserved (i.e. was \code{NA}). Inherits from \code{weightedAdj} and the previous \code{scanList} class (theoretical
or empirical, inheriting from \code{scanList}), and keeps track of the \code{scan.list}'s list of
attributes \code{attrs}.
}
\description{
Written as a S3 method to be applied to \code{scanList} or \code{sLlist} (list of \code{scanList}) objects.
}
\examples{
set.seed(42)
n <- 5L
samp.effort <- 241L

# Adjacency matrix import
## random directed adjacency matrix
Adj <- sample(1:samp.effort,n * n) |>
  matrix(nrow = 5,dimnames = list(letters[1:n],letters[1:n]))
diag(Adj) <- 0L
Adj

# social network simulations
## theoretical scans
sL <- simunet(Adj = Adj,samp.effort = samp.effort,mode = "directed",n.scans = 120L)
sL

## group-scan sampling
sL |> perform_exp(design_exp(customize_sampling("group",.6))) |> count_NA()
}
\seealso{
\code{\link[=simunet]{simunet()}}, \code{\link[=design_exp]{design_exp()}}, \code{\link[=perform_exp]{perform_exp()}}, \code{\link[=count_nonNA]{count_nonNA()}}.
}
