% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_tools.R
\name{design_exp}
\alias{design_exp}
\title{TO WRITE}
\usage{
design_exp(..., .dir = c("forward", "backward"))
}
\arguments{
\item{...}{TO WRITE}

\item{.dir}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
