% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_sampling.R
\name{determine_obsProb}
\alias{determine_obsProb}
\title{TO WRITE}
\usage{
determine_obsProb(
  scan.list,
  sampling = c("constant", "matrix", "random", "function"),
  all.sampled
)
}
\arguments{
\item{scan.list}{TO WRITE}

\item{sampling}{TO WRITE}

\item{all.sampled}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\keyword{internal}
