% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.tools.R
\name{diagonal}
\alias{diagonal}
\title{Diagonal Part of a Matrix}
\usage{
diagonal(M, output = c("matrix.logical", "vector.values"))
}
\arguments{
\item{M}{matrix or other R object with length(dim(x)) == 2. For back compatibility reasons, when the above is not fulfilled, as.matrix(x) is called first.}

\item{output}{choose to either return a matrix of logical values (TRUEs on the diagonal, i.e. comparable behaviour than diag() for instance) or a vector of the subsetted values of the matrix.}
}
\value{
square logical matrix with diagonal of TRUEs
}
\description{
Similarly to upper.tri and lower.tri, returns a matrix of logicals to identify the diagonal of a square matrix
}
\examples{
M<- matrix(sample(1:10,16,replace = TRUE),4,4)
diagonal<- function(M) {upper.tri(M,diag = TRUE)&!upper.tri(M,diag = FALSE)}
diagonal(M)

}
