% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_sampling.R
\name{draw_focalList}
\alias{draw_focalList}
\title{TO WRITE}
\usage{
draw_focalList(
  scan.list,
  sampling = c("even", "random", "function"),
  all.sampled = TRUE
)
}
\arguments{
\item{scan.list}{TO WRITE}

\item{sampling}{TO WRITE}

\item{all.sampled}{TO WRITE}

\item{sampling_fun}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\keyword{internal}
