% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeProb.R
\name{generate_edgeProb}
\alias{generate_edgeProb}
\title{\code{edgeProb} object generator}
\usage{
generate_edgeProb(Adj, samp.effort, mode, Adj.subfun = NULL)
}
\arguments{
\item{Adj}{integer matrix, the adjacency matrix (see \code{\link[=simunet]{simunet()}})}

\item{samp.effort}{integer scalar, the sampling effort (see \code{\link[=simunet]{simunet()}})}

\item{mode}{character scalar, the network igraph's mode (see \code{\link[=simunet]{simunet()}})}

\item{Adj.subfun}{function, the matrix subsetting function relevant for the adjacency matrix
\code{mode} (see \code{\link[=simunet]{simunet()}})}
}
\value{
an \code{edgeProb} object, i.e. a list containing:
\itemize{
\item \code{P}: the edge presence probability matrix
\item \code{Adj}: the inputted \code{Adj}
\item \code{samp.effort}: the inputted \code{samp.effort}
\item \code{mode}: the inputted \code{mode}
\item \code{Adj.subfun}: the inputted \code{Adj.subfun}
}
}
\description{
\code{edgeProb} object generator
}
\keyword{internal}
