% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeProb.R
\name{generate_edgeProb}
\alias{generate_edgeProb}
\title{TO WRITE}
\usage{
generate_edgeProb(Adj, samp.effort, mode, Adj.subfun = NULL)
}
\arguments{
\item{Adj}{TO WRITE}

\item{samp.effort}{TO WRITE}

\item{mode}{TO WRITE}

\item{Adj.subfun}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
