% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_tools.R
\name{generate_expDesign}
\alias{generate_expDesign}
\title{Generate \code{expDesign} objects}
\usage{
generate_expDesign(FUN.seq, fun.input)
}
\arguments{
\item{FUN.seq}{function sequence created by \code{purrr}'s \code{\link[purrr:compose]{compose()}} function}

\item{fun.input}{functions of \code{scanList} objects. Can be a combination of the provided "building
blocks" and user-defined function (cf. the example section).}
}
\value{
an \code{expDesign} objects, a list consisting in the following elements:
\itemize{
\item \code{FUN.seq}: function sequence created by \code{purrr}'s \code{\link[purrr:compose]{compose()}} function
\item WIP: more to come, notably to include more descriptive function names.
}
}
\description{
Generate \code{expDesign} objects
}
\keyword{internal}
