% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal.R
\name{generate_focal}
\alias{generate_focal}
\title{Generator for `focal` objects}
\usage{
generate_focal(focal.list, scan.number)
}
\arguments{
\item{focal.list}{a `focalList` object}

\item{scan.number}{either:
 \itemize{
  \item{an integer vector included in `1:total_scan` of the scans to perform}
  \item{the special case `"all"` (default) sets `scan.number` to `1:total_scan` and set the simulation to perform all the scans}
}}
}
\value{
an `focal` object (S3 class) containing:
\itemize{
  \item{`focal`: named integer vector representing the _index_ of the node(s) to sample at scan number(s) `scan.number`}
  \item{`focal.list`: inputted `focal.list` object}
  \item{`scan.number`:  inputted `scan.number`. If `"all"` was inputted, is set to `1:total_scan`}
}
}
\description{
Generator for `focal` objects
}
\examples{

set.seed(42)

n<- 5;nodes<- letters[1:n];total_scan<- 42;
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[non.diagonal(Adj)]<- sample(0:total_scan,n*(n-1),replace = TRUE)
Adj

focal.list<- generate_focalList(Adj,total_scan,focal.prob_fun = "even")
generate_focal(focal.list,5)
generate_focal(focal.list,3:10)
generate_focal(focal.list,"all")
# below returns an error if `scan.number` is incompatible with `total_scan`
# generate_focal(focal.list,50)
# generate_focal(focal.list,40:43)

}
