% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_sampling.R
\name{mask_non.focals}
\alias{mask_non.focals}
\title{TO WRITE}
\usage{
mask_non.focals(scan.list, focalList)
}
\arguments{
\item{scan.list}{TO WRITE}

\item{focalList}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\keyword{internal}
