% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_building.blocks.R
\name{remove_mostPeripheral}
\alias{remove_mostPeripheral}
\title{Remove from all scans the (overall) most peripheral individual
Individual centrality based on eigen vectors. Mostly given as an example of experimental
manipulations that could be performed on \code{scanList} as \code{expDesign}, even as user-defined
functions}
\usage{
remove_mostPeripheral(scan.list, ...)
}
\arguments{
\item{scan.list}{a \code{scanList} or \code{sLlist} object. See objects returned by
\code{\link[=simunet]{simunet()}}}

\item{...}{additional arguments to be passed. At the moment \code{scale_scans()} does not use

At the moment \code{remove_mostPeripheral()} does not use additional argument, arguments passed will be ignored.}
}
\value{
a \code{scanList}, or list of such, in which the most peripheral node has been removed. See
also \code{\link[=simunet]{simunet()}}.
}
\description{
Written as a S3 method to be applied to \code{scanList} or \code{sLlist} (list of \code{scanList}) objects.
}
\examples{
set.seed(42)
n <- 5L
samp.effort <- 241L

# Adjacency matrix import
## random directed adjacency matrix
Adj <- sample(1:samp.effort,n * n) |>
  matrix(nrow = 5,dimnames = list(letters[1:n],letters[1:n]))
Adj[lower.tri(Adj,diag = FALSE)] <- 0L
Adj

# social network simulations
## theoretical scans
sL <- simunet(Adj = Adj,samp.effort = samp.effort,mode = "upper",n.scans = 120L)
sL

## focal-scan sampling
sL |> perform_exp(design_sampling("focal","even")) |> remove_mostPeripheral()
}
\seealso{
\code{\link[=simunet]{simunet()}}, \code{\link[=design_exp]{design_exp()}}, \code{\link[=perform_exp]{perform_exp()}}.
}
