% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simunet.R
\name{simunet}
\alias{simunet}
\title{TO WRITE}
\usage{
simunet(
  Adj = NULL,
  samp.effort = NULL,
  mode = c("directed", "undirected", "max", "min", "upper", "lower", "plus", "vector"),
  n.scans = NULL,
  exp.design = NULL,
  ...,
  edge.Prob = NULL
)
}
\arguments{
\item{Adj}{TO WRITE}

\item{samp.effort}{TO WRITE}

\item{mode}{TO WRITE}

\item{n.scans}{TO WRITE}

\item{exp.design}{TO WRITE}

\item{...}{TO WRITE}

\item{edge.Prob}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
