% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_tools.R
\name{sLapply}
\alias{sLapply}
\title{TO WRITE}
\usage{
sLapply(sL, .f, ..., USE.NAMES = TRUE)
}
\arguments{
\item{sL}{TO WRITE}

\item{.f}{TO WRITE}

\item{...}{TO WRITE}

\item{USE.NAMES}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
\keyword{internal}
