% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanList_tools.R
\name{sLlapply}
\alias{sLlapply}
\title{TO WRITE}
\usage{
sLlapply(sLlist, FUN, ...)
}
\arguments{
\item{sLlist}{TO WRITE}

\item{FUN}{TO WRITE}

\item{...}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
